feat: add fixed64 runtime
